fix: resolve Windows/MSVC build failures (D8021 invalid flag + C4244 narrowing warnings)
Agent-Logs-Url: https://github.com/AngelzXHub/Fission/sessions/ffb9837f-1952-4a98-9307-c7170393b304

+++ Fission.Decompiler/include/ControlFlowAnalyzer.hpp
@@ -102,7 +102,7 @@ struct SSARef {
 
     bool operator==(const SSARef &other) const { return regIndex == other.regIndex && version == other.version; }
     SSARef() = default;
-    SSARef(int32_t reg, int32_t ver) : regIndex(reg), version(ver) {}
+    SSARef(int32_t reg, int32_t ver) : regIndex(static_cast<uint8_t>(reg)), version(ver) {}
 };
 
 namespace std {
